Fix all remaining problems that made rendering in the meter scale glitchy taking off from mars also glitchy.

+++ sp/src/star.cpp
@@ -47,7 +47,7 @@ void CStar::PostRender(bool bTransparent) const
 	c.SetColor(Color(255, 255, 255, 255));
 	c.SetLighting(false);
 
-	c.RenderBillboard("textures/star-yellow.png", (GetRenderScalableRadius()*2).GetUnits(SPGame()->GetSPRenderer()->GetRenderingScale()));
+	c.RenderBillboard("textures/star-yellow.png", (float)(GetRenderScalableRadius()*2).GetUnits(SPGame()->GetSPRenderer()->GetRenderingScale()));
 }
 
 CScalableFloat CStar::GetCloseOrbit()
